+++ apps/cluster_manager/config/dev.exs
@@ -27,6 +27,4 @@ use Mix.Config
 # Configuration from the imported file will override the ones defined
 # here (which is why it is important to import them last).
 #
-config :cluster_manager, cluster_status: false
-
-import_config "#{Mix.env()}.exs"
+#     import_config "#{Mix.env()}.exs"