+++ apps/cluster_manager/test/cluster_manager_test.exs
@@ -1,8 +1,4 @@
 defmodule ClusterManagerTest do
   use ExUnit.Case
   doctest ClusterManager
-
-  test "greets the world" do
-    assert ClusterManager.hello() == :world
-  end
 end