+++ apps/game_services_web/lib/game_services_web/application.ex
@@ -6,8 +6,6 @@ defmodule GameServicesWeb.Application do
   use Application
 
   def start(_type, _args) do
-
-
     # List all child processes to be supervised
     children = [
       # Start the endpoint when the application starts